pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "voicemon"
version = "0.1.0"
description = "Production observability platform for Voice AI agents — turn-level tracing, latency waterfalls, anomaly detection, and alerting across LiveKit, Pipecat, and Vapi."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [{ name = "VoiceMon Team" }]
keywords = ["voice-ai", "observability", "monitoring", "livekit", "pipecat", "vapi", "opentelemetry"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
]

dependencies = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "redis>=5.0.0",
    "asyncpg>=0.29.0",
    "prometheus-client>=0.20.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
livekit = ["livekit-agents>=0.12.0"]
pipecat = ["pipecat-ai>=0.0.50"]
vapi = ["fastapi>=0.110.0", "uvicorn>=0.27.0"]
dashboard = ["streamlit>=1.30.0", "plotly>=5.18.0", "pandas>=2.1.0"]
all = [
    "livekit-agents>=0.12.0",
    "pipecat-ai>=0.0.50",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "streamlit>=1.30.0",
    "plotly>=5.18.0",
    "pandas>=2.1.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.urls]
Homepage = "https://github.com/voicemon/voicemon"
Documentation = "https://voicemon.dev"
Repository = "https://github.com/voicemon/voicemon"

[tool.hatch.build.targets.wheel]
packages = ["voicemon"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
